Fix the scrolling to bottom issue temporarily using asyncAfter method.
